Rearrange stores that're specific to the draw page
