+++ src/lib/draw/polygon/PolygonControls.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
+  import { polygonTool } from "lib/draw/stores";
   import { SecondaryButton } from "lib/govuk";
-  import { polygonTool } from "stores";
   import { undoLength } from "./stores";
 
   function undo() {
