+++ src/lib/draw/polygon/PolygonMode.svelte
@@ -1,9 +1,14 @@
 <script lang="ts">
   import type { Polygon } from "geojson";
+  import {
+    gjSchemeCollection,
+    mode,
+    newFeatureId,
+    polygonTool,
+  } from "lib/draw/stores";
   import { ButtonGroup, DefaultButton, SecondaryButton } from "lib/govuk";
   import type { FeatureWithProps } from "lib/maplibre";
   import { getArbitraryScheme } from "lib/sidebar/scheme_data";
-  import { gjSchemeCollection, mode, newFeatureId, polygonTool } from "stores";
   import { onDestroy, onMount } from "svelte";
   import type { Feature } from "types";
   import PolygonControls from "./PolygonControls.svelte";
