+++ src/lib/draw/route/RouteSnapperLoader.svelte
@@ -1,7 +1,8 @@
 <script lang="ts">
+  import { routeTool } from "lib/draw/stores";
   import { ErrorMessage } from "lib/govuk";
   import init from "route-snapper";
-  import { map, routeTool } from "stores";
+  import { map } from "stores";
   import { onMount } from "svelte";
   import { RouteTool } from "./route_tool";
