+++ src/lib/forms/PipelineForm.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+  import { gjSchemeCollection, routeTool } from "lib/draw/stores";
   import {
     Checkbox,
     FormElement,
@@ -8,7 +9,6 @@
     TextArea,
   } from "lib/govuk";
   import { prettyPrintMeters } from "lib/maplibre";
-  import { gjSchemeCollection, routeTool } from "stores";
   import type { InterventionProps } from "types";
   import PipelineType from "./PipelineType.svelte";
