+++ src/lib/sidebar/ListMode.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { FileInput } from "lib/common";
+  import { gjSchemeCollection } from "lib/draw/stores";
   import { ErrorMessage, SecondaryButton } from "lib/govuk";
-  import { gjSchemeCollection } from "stores";
   import PerSchemeControls from "./PerSchemeControls.svelte";
   import { addEmptyScheme, backfill } from "./scheme_data";
